+++ server/controllers/animalController.js
@@ -1,5 +1,6 @@
 const Animal = require('../models/AnimalModel');
 
+
 // get all animals
 const getAllAnimals = async (req, res) => {
     try {
@@ -25,18 +26,6 @@ const getSingleAnimal = async (req, res) => {
         res.status(400).json({ message: 'Error getting animal', err });
     }
 };
-
-// create a new animal
-const createAnimal = async (req, res) => {
-    const {animal_id, name, age, sex, animal_type, images_and_videos, description, contact_user, area_of_adoption, color} = req.body;
-
-    try {
-        const animal = await Animal.create({animal_id, name, age, sex, animal_type, images_and_videos, description, contact_user, area_of_adoption, color});
-        res.status(200).json({animal});
-    } catch (err) {
-        res.status(400).json({mssg: 'error creating animal', err})
-    }
-}
 
 
 // delete a animal
@@ -71,7 +60,6 @@ const updateAnimal = async (req, res) => {
 };
 
 module.exports = {
-    createAnimal,
     getAllAnimals,
     getSingleAnimal,
     deleteAnimal,
